feat: add unauthenticated ingress relation endpoint (#60)
Previously, if the ingress was configured with authentication then all ingressed apps (apps related to `ingress`) had authentication applied to them.  This meant we could never have some applications ingressed with auth alongside others without auth.

This commit implements a second ingress relation, `ingress-unauthenticated`, which allows for applications to request an ingressed route that will always be unauthenticated (even if the ingress is configured with authentication).  This allows for some applications to be authenticated (say a private dashboard) while others are unauthenticated (a /status endpoint indicating that the system is up).